refactor(typing): move subclassing types to own file
second part of splitting the files, necessary to decomple types interdependencies caused by
inferring large portion of types based on the hook definitions
