+++ src/types/subclassing.ts
@@ -6,68 +6,10 @@ import theWebpack from 'webpack';
 import type { Ploadin } from 'ploadin';
 import type { Tapable } from 'tapable';
 
-import type { Constructor, AnyFunc, RequiredKeys } from './util';
+import type { RequiredKeys } from './util';
 import type { Overrides, ActiveHooks } from './hook';
-import type { ModuleBase, DependencyBase } from './base';
-import type { ModuleFilename } from './module-filename';
 
 export { ParamsDefault } from './subclassing-params';
-
-export interface DependencyTemplate {
-  apply: AnyFunc;
-}
-
-export type DependencyTemplateClass<
-  T extends DependencyTemplate = DependencyTemplate
-> = Constructor<T> & typeof Tapable;
-
-export interface Dependency
-  extends theWebpack.compilation.Dependency,
-    DependencyBase {
-  identifierIndex: number;
-}
-
-export type DependencyClass<T extends Dependency = Dependency> = Constructor<
-  T
-> & {
-  compare: typeof theWebpack.compilation.Dependency['compare'];
-};
-
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface Module extends ModuleBase {}
-
-export type ModuleClass<T extends Module = Module> = Constructor<T>;
-
-export type ModuleFactoryData<D extends Dependency = Dependency> = {
-  dependencies: D[];
-} & { [key: string]: any };
-
-export type ModuleFactoryCallback<M extends Module = Module> = (
-  error: Error | null,
-  result: M,
-) => void;
-
-export interface ModuleFactory<
-  // Allow user to specify the types by passing an object of types as params
-  T extends {
-    dependency?: Dependency;
-    module?: Module;
-  } = {},
-  // Set defaults
-  Dep extends Dependency = T['dependency'] extends Dependency
-    ? T['dependency']
-    : Dependency,
-  Mod extends Module = T['module'] extends Module ? T['module'] : Module
-> {
-  create(
-    data: ModuleFactoryData<Dep>,
-    callback: ModuleFactoryCallback<Mod>,
-  ): void;
-}
-
-export type ModuleFactoryClass<
-  T extends ModuleFactory = ModuleFactory
-> = Constructor<T> & typeof Tapable;
 
 /**
  * Options passed to class factory.
@@ -109,16 +51,6 @@ export interface ClassOptions<T extends ParamsDefault = {}>
   moduleClass?: ModCls;
   dependencyTemplateClass?: DepTemplateCls;
 }
-
-/**
- * Default options passed to MiniExtractPlugin constructor
- */
-export type ConstructorOptions = {
-  filename?: string;
-  moduleFilename?: ModuleFilename;
-  chunkFilename?: string;
-  ignoreOrder?: boolean;
-};
 
 /**
  * MiniExtractPlugin instance
